PHP conflicts

+++ inc/thisyear.php
@@ -45,7 +45,8 @@ function lwtv_yikes_this_year_dead( $thisyear ) {
 			// Jesus, I hope no one dies twice in the same year ... SARA
 			$died_date = get_post_meta( $dead_char, 'lezchars_death_year', true );
 			foreach ( $died_date as $date ) {
-				if ( (int) substr( $date, 0, 4 ) === substr( $date, 0, 4 ) ) {
+
+				if ( (int) substr( $date, 0, 4 ) === (int) substr( $date, 0, 4 ) ) {
 					$died_year  = substr( $date, 0, 4 );
 					$died_array = date_parse_from_format( 'Y-m-d', $date );
 				} else {
